+++ Phoenix Editor/src/Application.h
@@ -8,6 +8,11 @@
 
 namespace Phoenix{
 
+	struct EditorConfig{
+		bool VSync = false;
+		std::string last_opened_project;
+	};
+
 	
 	class Editor : public Engine{
 		public:
@@ -18,12 +23,14 @@ namespace Phoenix{
 			void create() override;
 			void render() override;
 
-
 			void runProject();
+
+			std::string saveConfig();
 
 		public:
 			RendererImGui renderer_ImGui;
 			Project project;
+			EditorConfig config;
 
 
 		private:
@@ -32,6 +39,7 @@ namespace Phoenix{
 			void zoom_camera(float dy);
 			void place_camera();
 			void use_editor_camera();
+
 
 		private:
 			std::string _path;
@@ -43,6 +51,7 @@ namespace Phoenix{
 
 			int _running_counter = 0;
 			bool _can_render = true;
+
 			
 	};